# Benchmark `count` for `vector<bool>`