--- a/benchmarks/src/vector_bool_count.cpp
+++ b/benchmarks/src/vector_bool_count.cpp
@@ -11,10 +11,10 @@
 using namespace std;
 
 vector<bool> createRandomVector(const size_t size) {
-    static mt19937 gen;
-    static bernoulli_distribution dist{0.5};
+    mt19937 gen;
+    bernoulli_distribution dist{0.5};
     vector<bool> result(size);
-    generate(result.begin(), result.end(), [] { return dist(gen); });
+    generate(result.begin(), result.end(), [&] { return dist(gen); });
     return result;
 }