@StephanTLavavej 2025-08-22
5950X results:

Benchmark           |  Before |   After | Speedup
--------------------|---------|---------|--------
count_aligned/64    | 18.8 ns | 18.8 ns | 1.00
count_aligned/4096  | 77.7 ns | 62.8 ns | 1.24
count_aligned/65536 |  903 ns |  786 ns | 1.15